pyproject.toml
[tool.poetry]
name = "demat-folio-api"
version = "0.1.0"
description = ""
authors = ["Nitheesh-S <nitheeshmsk@gmail.com>"]
readme = "README.md"
packages = [{include = "demat_folio_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
fyers-apiv3 = "^3.0.3"
python-decouple = "^3.8"
django-ses = "^3.5.0"
cryptography = "^41.0.4"
kiteconnect = "^4.2.0"
googlefinance = "^0.7"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
black = "^23.9.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
